+++ packages/gamut-labs/src/GlobalHeader/types.tsx
@@ -1,6 +1,9 @@
 import { ReactNode } from 'react';
 
-import { AppHeaderClickHandler } from '../AppHeader/AppHeaderElements/types';
+import {
+  AppHeaderClickHandler,
+  AppHeaderItemWithHref,
+} from '../AppHeader/AppHeaderElements/types';
 import { AppHeaderSearch } from '../AppHeader/Search/useHeaderSearch';
 
 type RenderNotifications = {
@@ -13,7 +16,7 @@ type BaseHeader = {
   /** A method to be called on click/activating a header item */
   action: AppHeaderClickHandler;
   /** A method to be called only on click/activating a *link* header item */
-  onLinkAction?: AppHeaderClickHandler;
+  onLinkAction?: AppHeaderClickHandler<AppHeaderItemWithHref>;
   className?: string;
   search: AppHeaderSearch;
 };